Add files via upload
REQUIREMENTS :

These requirements are rather high-level and vague. If details are omitted, it is because we will be happy with any of a wide variety of solutions. Don't worry about finding "the" solution. Feel free to be creative with the requirements. Your goal is to impress (but do so with clean code).

Create a browser based or native app to provide information on NYC High schools.

    Display a list of NYC High Schools.
    Get your data here: https://data.cityofnewyork.us/Education/DOE-High-School-Directory-2017/s3k6-pzi2
    Selecting a school should show additional information about the school
    Display all the SAT scores - include Math, Reading and Writing.
    SAT data here: https://data.cityofnewyork.us/Education/SAT-Results/f9bf-2cp4
    It is up to you to decide what additional information to display

When creating a name for your project, please use the following naming convention:

YYYYMMDD-[First&LastName]-NYCSchools (Example: 20180101-DanielleBordner-NYCSchools)

In order to prevent you from running down rabbit holes that are less important to us, try to prioritize the following: